colorchecker_requirements.txt
# Requirements for ColorChecker detection scripts

# Core dependencies
opencv-contrib-python>=4.8.0  # Includes ArUco module
numpy>=1.24.0
Pillow>=10.0.0

# PDF generation
reportlab>=4.0.0

# Optional: for advanced features
# scipy>=1.11.0  # For advanced color processing
